requirements.txt
# Pan-African Microfinance SaaS — Python Dependencies
# Core
Django==5.1.*
djangorestframework==3.15.*
django-cors-headers==4.*
django-filter==24.*
psycopg2-binary==2.9.*
gunicorn==23.*

# Task queue
celery==5.4.*
django-celery-beat==2.7.*
redis==5.*

# Authentication
PyJWT==2.9.*
cryptography==43.*

# Supabase client (for Storage and direct calls)
supabase==2.*
httpx==0.27.*

# PDF generation
weasyprint==62.*
# Excel export
openpyxl==3.1.*

# SMS
africastalking==1.2.*

# Email
resend==2.*

# Utilities
python-decouple==3.8
whitenoise==6.*

# Dev/testing
pytest-django==4.*
factory-boy==3.*
faker==30.*
